ward,location and sub-locations data stored and fetched from the database
