+++ students/deletedapp.php
@@ -724,3 +724,863 @@ elseif($kilibwoni_sub_location  == '' && $kapkangani_sub_location  == '' && $lol
 && empty($_POST['kapkangani_sub_location']) && empty($_POST['lolminingai_sub_location']) && empty($_POST['kipsigak_sub_location'])&& empty($_POST['tulon_sub_location'])
   && empty($_POST['kipture_sub_location']) && empty($_POST['kabirirsang_sub_location']) && empty($_POST['arwos_sub_location']) && empty($_POST['terige_sub_location']) && empty($_POST['kaplamai_sub_location'])
  && empty($_POST['kamobo_sub_location']) && empty($_POST['township_sub_location']) && empty($_POST['kiminda_sub_location']) && empty($_POST['chepkumia_sub_location'])
+
+
+<option value="Chepkumia">Chepkumia</option>
+<option value="Kapkangani">Kapkangani</option>
+<option value="Kilibwoni">Kilibwoni</option>
+<!--  -->
+function showKap(optionValue) {
+      const secondSelect = document.getElementById('sec');
+      secondSelect.innerHTML = ''; // Clear the existing options
+
+      const optionData = {
+        Kapkangani: ['','Chepsonoi', 'Tindinyo', 'Kiborgok'],
+      };
+
+      if (optionData[optionValue]) {
+        optionData[optionValue].forEach(option => {
+          const optionElement = document.createElement('option');
+          optionElement.value = option;
+          optionElement.textContent = option;
+          secondSelect.appendChild(optionElement);
+        });
+      }
+}
+function showChep(optionValue) {
+      const secondSelect = document.getElementById('sec');
+      secondSelect.innerHTML = ''; // Clear the existing options
+
+      const optionData = {
+        Chepkumia: ['','Chepkumia', 'Cheboite'],
+      };
+
+      if (optionData[optionValue]) {
+        optionData[optionValue].forEach(option => {
+          const optionElement = document.createElement('option');
+          optionElement.value = option;
+          optionElement.textContent = option;
+          secondSelect.appendChild(optionElement);
+        });
+      }
+}
+function showO(optionValue){
+    const secondSelect = document.getElementById('sec');
+      secondSelect.innerHTML = ''; // Clear the existing options
+
+      const optionData = {
+        Kamobo: ['','Kamobo'],
+        Township: ['','Township'],
+        Kiminda: ['','Kiminda', 'Meswo'],
+      };
+
+      if (optionData[optionValue]) {
+        optionData[optionValue].forEach(option => {
+          const optionElement = document.createElement('option');
+          optionElement.value = option;
+          optionElement.textContent = option;
+          secondSelect.appendChild(optionElement);
+        });
+      }
+}
+
+function showK(optionValue) {
+      const secondSelect = document.getElementById('sec');
+      secondSelect.innerHTML = ''; // Clear the existing options
+
+      const optionData = {
+        Kilibwoni: ['','Kilibwoni', 'Kapnyerebai', 'Kaplonyo'],
+        Lolminingai: ['','Lolminingai','Kabore', 'Ndubeneti', 'Kaplolok'],
+        Kipsigak: ['','Kipsotoi', 'Kipsigak','Kakeruge'],
+        Kipture: ['','Kipture', 'Kimaam', 'Irimis'],
+        Kabirirsang: ['','Kabirirsang','Underit','Chesuwe'],
+        Arwos: ['','Tiryo','Arwos','Kaptendon'],
+        Kaplamai: ['','Kaptangunyo','Kapchemoiywo','Kaplamai','Kapsumbeiywo'],
+        Tulon: ['','Kapchumba','Tulon','Kapchepsir','Kiptenden'],
+        Terige: ['','Song`oliet','Terige','Emdin'],
+      };
+
+      if (optionData[optionValue]) {
+        optionData[optionValue].forEach(option => {
+          const optionElement = document.createElement('option');
+          optionElement.value = option;
+          optionElement.textContent = option;
+          secondSelect.appendChild(optionElement);
+        });
+      }
+    }
+<!-- end of the functions -->
+
+// starts here
+
+// $(document).ready(function() {
+//   // Get the progress bar element.
+//   const progressBarElement = $('.progress-bar');
+
+//   // Get the input field elements.
+//   const input1 = $('#first');
+//   const input2 = $('#last');
+//   const input3 = $('#gender');
+//   const input4 = $('#date');
+// //   const input5 = $('#parent');
+//   const input5 = $('#phone');
+//   const input6 = $('#id');
+// //   const input8 = $('#email');
+// //   const input9 = $('#county');
+//   const input7 = $('#opts');
+// //   const input11 = $('#default');
+//   const input8 = $('#subs');
+//   const input9 = $('#level');
+//   const input10 = $('#reg');
+// //   const input15= $('#soValue');
+//   const input11= $('#file-upload-field');
+//   const input12= $('#file-upload');
+// //   const input18= $('#year');
+  
+
+
+//   // Calculate the initial percentage.
+//   let percentage = 0;
+  
+
+//   // Get the number of input fields.
+//   const numberOfInputFields = 12;
+
+//   // Check if any of the input fields are already filled.
+//   if (input1.val() || input2.val() || input3.val()|| input4.val()|| input5.val()|| input6.val()|| input7.val()|| input8.val()|| input9.val()|| input10.val()
+//   || input11.val()|| input12.val()) {
+//     // If any of the input fields are already filled, calculate the percentage based on the number of filled input fields.
+//     percentage = (100 / numberOfInputFields) * (input1.val() ? 1 : 0) + (input2.val() ? 1 : 0) + (input3.val() ? 1 : 0)+ (input4.val() ? 1 : 0)+ (input5.val() ? 1 : 0)
+//     + (input6.val() ? 1 : 0)+ (input7.val() ? 1 : 0)+ (input8.val() ? 1 : 0)+ (input9.val() ? 1 : 0)+ (input10.val() ? 1 : 0)+ (input11.val() ? 1 : 0)+ (input12.val() ? 1 : 0)
+//     ;
+//   }
+
+//     var completedFields = 0;
+//   // Update the progress bar percentage.
+//   progressBarElement.css('width', percentage + '%');
+//   $('.percentage').text(percentage + '%');
+
+//   // Add an event listener for the change event on the input field elements.
+//   input1.change(function() {
+//        if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+     
+//   });
+
+//   input2.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//      if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+
+//   input3.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//    if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input4.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//      if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input5.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//      if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input6.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//     percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//      if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input7.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//     if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input8.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//      if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input9.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//     if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input10.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//    if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input11.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//    if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+//   input12.change(function() {
+//     // Calculate the new percentage.
+//     // percentage = percentage + (100 / numberOfInputFields);
+
+//     // Update the progress bar percentage.
+//    if($(this).val() !== ''){
+//             percentage = Math.floor(percentage + (100 / numberOfInputFields));
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }else{
+//             percentage = Math.floor(percentage - (100 / numberOfInputFields) +1);
+//     progressBarElement.css('width', percentage + '%');
+//     $('.percentage').text(percentage + '%');
+//         }
+//   });
+  
+// });
+
+// const inputFields = document.querySelectorAll('input[required]');
+// const progressBar =document.querySelector('.progress');
+// inputFields.forEach(input => {
+//     input.addEventListener('input', updateProgressBar);
+// });
+// function updateProgressBar(){
+//     const completedFields = document.querySelectorAll('input[required]:not(:empty)');
+//     const progress =completedFields.length / inputFields.length * 100;
+//     progressBar.style.width = '${progress}%';
+// }
+
+// $(document).ready(function(){
+//     var totalFields = $('input[type="text"]').length;
+//     var completedFields = 0;
+//     $('input[type="text"]').change(function(){
+//         if($(this).val() !== ''){
+//             completedFields++;
+//         }else{
+//             completedFields--;
+//         }
+
+//         var percentage = Math.floor((completedFields / totalFields) * 100);
+//         $('.progress').css('width', percentage + '%');
+//         $('.percentage').text(percentage + '%');
+//     });
+// });
+
+<!-- sub_location scripts -->
+function showOptions() {
+  
+  var parentValue = document.getElementById("kapkangani").value;
+  if(parentValue === 'Kapkangani'){
+      document.getElementById('kapkangani_sub1').style.display = 'block';
+      document.getElementById('kapkangani_sub2').style.display = 'block';
+      document.getElementById('kapkangani_sub3').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+
+  }else if(parentValue === ''){
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('sub').style.display = 'block';
+  }
+  
+}
+
+function Options() {
+  var parentV = document.getElementById("chepkumia").value;
+if(parentV === "Chepkumia"){
+      document.getElementById('chepkumia_sub1').style.display = 'block';
+      document.getElementById('chepkumia_sub2').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+
+      }else if(parentV === ''){
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('sub').style.display = 'block';
+      }
+
+}
+
+function show() {
+  var parentVa = document.getElementById("kilibwoni").value;
+if(parentVa === "Kilibwoni"){
+      document.getElementById('kilibwoni_sub1').style.display = 'block';
+      document.getElementById('kilibwoni_sub2').style.display = 'block';
+      document.getElementById('kilibwoni_sub3').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+
+      }else if(parentVa === 'Lolminingai'){
+      document.getElementById('lolminingai_sub1').style.display = 'block';
+      document.getElementById('lolminingai_sub2').style.display = 'block';
+      document.getElementById('lolminingai_sub3').style.display = 'block';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub1').style.display = 'none';
+      document.getElementById('kapkangani_sub2').style.display = 'none';
+      document.getElementById('kapkangani_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub').style.display = 'none';
+      document.getElementById('chepkumia_sub1').style.display = 'none';
+      document.getElementById('chepkumia_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('kipture_sub').style.display = 'none';
+      document.getElementById('kabirirsang_sub').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub').style.display = 'none';
+      }else if(parentVa === 'Kipsigak'){
+      document.getElementById('kipsigak_sub1').style.display = 'block';
+      document.getElementById('kipsigak_sub2').style.display = 'block';
+      document.getElementById('kipsigak_sub3').style.display = 'block';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }else if(parentVa === 'Kipture'){
+      document.getElementById('kipture_sub1').style.display = 'block';
+      document.getElementById('kipture_sub2').style.display = 'block';
+      document.getElementById('kipture_sub3').style.display = 'block';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }
+      else if(parentVa === 'Kabirirsang'){
+      document.getElementById('kabirirsang_sub1').style.display = 'block';
+      document.getElementById('kabirirsang_sub2').style.display = 'block';
+      document.getElementById('kabirirsang_sub3').style.display = 'block';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }else if(parentVa === 'Arwos'){
+      document.getElementById('arwos_sub').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }else if(parentVa === 'Kaplamai'){
+      document.getElementById('kaplamai_sub').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }else if(parentVa === 'Tulon'){
+      document.getElementById('tulon_sub').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }else if(parentVa === 'Terige'){
+      document.getElementById('terige_sub').style.display = 'block';
+      document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+      }
+      
+      else if(parentVa === ''){
+      document.getElementById('kilibwoni_sub').style.display = 'none';
+      document.getElementById('sub').style.display = 'block';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kipsigak_sub').style.display = 'none';
+      document.getElementById('kipture_sub').style.display = 'none';
+      document.getElementById('kabirirsang_sub').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      }
+
+}
+
+function showOpt(){
+  var parent = document.getElementById("kapsabet").value;
+if(parent === 'Kamobo'){
+  document.getElementById('kamobo_sub').style.display = 'block';
+  document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+
+}else if(parent === 'Kiminda'){
+  document.getElementById('kiminda_sub1').style.display = 'block';
+  document.getElementById('kiminda_sub2').style.display = 'block';
+  document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+      document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+
+}else if(parent === 'Township'){
+  document.getElementById('township_sub').style.display = 'block';
+  document.getElementById('kabirirsang_sub1').style.display = 'none';
+      document.getElementById('kabirirsang_sub2').style.display = 'none';
+      document.getElementById('kabirirsang_sub3').style.display = 'none';
+      document.getElementById('kipsigak_sub1').style.display = 'none';
+      document.getElementById('kipsigak_sub2').style.display = 'none';
+      document.getElementById('kipsigak_sub3').style.display = 'none';
+      document.getElementById('sub').style.display = 'none';
+      document.getElementById('main').style.display = 'none';
+      document.getElementById('kapkangani_sub').style.display = 'none';
+      document.getElementById('lolminingai_sub1').style.display = 'none';
+      document.getElementById('lolminingai_sub2').style.display = 'none';
+      document.getElementById('lolminingai_sub3').style.display = 'none';
+      document.getElementById('chepkumia_sub').style.display = 'none';
+      document.getElementById('kilibwoni_sub1').style.display = 'none';
+      document.getElementById('kilibwoni_sub2').style.display = 'none';
+      document.getElementById('kilibwoni_sub3').style.display = 'none';
+      document.getElementById('kipture_sub1').style.display = 'none';
+      document.getElementById('kipture_sub2').style.display = 'none';
+      document.getElementById('kipture_sub3').style.display = 'none';
+      document.getElementById('arwos_sub').style.display = 'none';
+      document.getElementById('kaplamai_sub').style.display = 'none';
+      document.getElementById('tulon_sub').style.display = 'none';
+      document.getElementById('terige_sub').style.display = 'none';
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub1').style.display = 'none';
+  document.getElementById('kiminda_sub2').style.display = 'none';
+  document.getElementById('chepkumia_sub1').style.display = 'none';
+  document.getElementById('chepkumia_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub1').style.display = 'none';
+  document.getElementById('kapkangani_sub2').style.display = 'none';
+  document.getElementById('kapkangani_sub3').style.display = 'none';
+
+}else if(parent === ''){
+  document.getElementById('township_sub').style.display = 'none';
+  document.getElementById('sub').style.display = 'block';
+  
+  document.getElementById('kamobo_sub').style.display = 'none';
+  document.getElementById('kiminda_sub').style.display = 'none';
+}
+}
